Refactoring CreateEvent
