+++ assets/data/VendorList.json
@@ -1,0 +1,94 @@
+[{
+    "_id": 1,
+    "image": "https://res.cloudinary.com/xenxei46/image/upload/v1682686746/photo_jockgx.png",
+    "title": "Photographer"
+  },{
+    "_id": {
+      "$oid": "5f15d92ee520d44421ed8e9b"
+    },
+    "image": "https://res.cloudinary.com/xenxei46/image/upload/v1682686741/dj-mixer_ubww35.png",
+    "title": "Disk Jockeys",
+    "category": {
+      "$oid": "5f15d5b7cb4a6642bddc0fe8"
+    },
+    "__v": 0
+  },{
+    "_id": {
+      "$oid": "5f15d964e520d44421ed8e9c"
+    },
+    "image": "https://res.cloudinary.com/xenxei46/image/upload/v1682686746/host_erxddt.png",
+    "title": "Hosts",
+    "category": {
+      "$oid": "5f15d5b7cb4a6642bddc0fe8"
+    },
+    "__v": 0
+  },
+  {
+    "_id": {
+      "$oid": "5f15d964e520d44421ed8e9c"
+    },
+    "image": "https://res.cloudinary.com/xenxei46/image/upload/v1682686746/laugh_agaht2.png",
+    "title": "Comedians",
+    "category": {
+      "$oid": "5f15d5b7cb4a6642bddc0fe8"
+    },
+    "__v": 0
+  },
+  {
+    "_id": {
+      "$oid": "5f15d964e520d44421ed8e9c"
+    },
+    "image": "https://res.cloudinary.com/xenxei46/image/upload/v1682686741/dancer_qszlh7.png",
+    "title": "Dancers",
+    "category": {
+      "$oid": "5f15d5b7cb4a6642bddc0fe8"
+    },
+    "__v": 0
+  },
+  {
+    "_id": {
+      "$oid": "5f15d964e520d44421ed8e9c"
+    },
+    "image": "https://res.cloudinary.com/xenxei46/image/upload/v1682686746/food_m2nbdr.png",
+    "title": "Cateres",
+    "category": {
+      "$oid": "5f15d5b7cb4a6642bddc0fe8"
+    },
+    "__v": 0
+  },
+  {
+    "_id": {
+      "$oid": "5f15d964e520d44421ed8e9c"
+    },
+    "image": "https://res.cloudinary.com/xenxei46/image/upload/v1682686746/musician_rfrqya.png",
+    "title": "Artiste",
+    "category": {
+      "$oid": "5f15d5b7cb4a6642bddc0fe8"
+    },
+    "__v": 0
+  },
+  {
+    "_id": {
+      "$oid": "5f15d964e520d44421ed8e9c"
+    },
+    "image": "https://res.cloudinary.com/xenxei46/image/upload/v1682686741/cinematographers_axkwoq.png",
+    "title": "cinematographers",
+    "category": {
+      "$oid": "5f15d5b7cb4a6642bddc0fe8"
+    },
+    "__v": 0
+  },
+  {
+    "_id": {
+      "$oid": "5f15d964e520d44421ed8e9c"
+    },
+    "image": "https://res.cloudinary.com/xenxei46/image/upload/v1682686741/boy_rexp5y.png",
+    "title": "Users",
+    "category": {
+      "$oid": "5f15d5b7cb4a6642bddc0fe8"
+    },
+    "__v": 0
+  }
+
+
+]
