+++ assets/data/vendors.json
@@ -2,7 +2,7 @@
   "_id": {
     "$oid": "5f15d8852a025143f9593a7c"
   },
-  "image": "https://res.cloudinary.com/xenxei46/image/upload/v1682521359/sandra_c7ez8u.png",
+  "image": "https://res.cloudinary.com/xenxei46/image/upload/v1682521174/cardi_uiwhel.png",
   "user": "Sandra Carson",
   "handle": "@scarzz",
   "career": "Photographer",
