+++ package.json
@@ -26,7 +26,7 @@
     "react": "18.2.0",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.43.7",
-    "react-native": "0.71.6",
+    "react-native": "0.71.7",
     "react-native-gesture-handler": "~2.9.0",
     "react-native-pager-view": "^6.1.2",
     "react-native-reanimated": "~2.14.4",
