+++ src/Screens/Dashboard/Categories/index.tsx
@@ -1,0 +1,14 @@
+import { StyleSheet, Text, View } from 'react-native'
+import React from 'react'
+
+const Categories = () => {
+  return (
+    <View>
+      <Text>Categories</Text>
+    </View>
+  )
+}
+
+export default Categories;
+
+const styles = StyleSheet.create({})
