+++ src/Screens/Dashboard/Categories/style.ts
@@ -1,0 +1,5 @@
+import { StyleSheet } from 'react-native'
+
+export const Styles = StyleSheet.create({
+    
+});
