--- src/Screens/Dashboard/stepperForm/style.ts
@@ -1,13 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-
-export const Styles = StyleSheet.create({
-    main_Container: {
-        width: '100%',
-        height: '100%',
-        // borderWidth: 2,
-        // borderStyle: 'solid',
-        backgroundColor: '#FCFCFC'
-    },
-    centerElement: {justifyContent: 'center', alignItems: 'center'}, 
-});
